requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
autopep8==1.5.7
botocore==1.29.139
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
Django==3.1
django-admin-thumbnails==0.2.5
django-auto-logout==0.5.0
django-ckeditor==5.7.1
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-phonenumber-field==6.0.0
django-robots==4.0
django-shortcuts==1.6
django-storages==1.11.1
django-taggit==1.5.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
phonenumbers==8.12.41
Pillow==9.0.0
psycopg2-binary==2.9.2
pycodestyle==2.7.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
python-memcached==1.59
pytz==2021.1
qrcode==6.1
reportlab==3.6.5
requests==2.25.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
validate-email==1.3
webencodings==0.5.1
xhtml2pdf==0.2.5
